requirements.txt
yfinance>=0.2.31
cassandra-driver>=3.29.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
flask>=3.0.0
